Move even more files around
