+++ front/app/containers/Admin/projectFolders/containers/projects/ItemsNotInFolder.tsx
@@ -5,7 +5,7 @@ import useAuthUser from 'api/me/useAuthUser';
 import { PublicationStatus } from 'api/projects/types';
 import useUpdateProjectFolderMembership from 'api/projects/useUpdateProjectFolderMembership';
 
-import ProjectRow from 'containers/Admin/projects/components/ProjectRow';
+import ProjectRow from 'containers/Admin/projects/_shared/components/ProjectRow';
 
 import { List, Row } from 'components/admin/ResourceList';
