+++ front/app/containers/Admin/projects/_shared/components/ProjectSetupForm/ProjectCardImageDropzone.tsx
