+++ front/app/containers/Admin/projects/_shared/components/ProjectSetupForm/ProjectCardImageTooltip.tsx
@@ -4,16 +4,16 @@ import { IconTooltip } from '@citizenlab/cl2-component-library';
 
 import { FormattedMessage, useIntl } from 'utils/cl-intl';
 
-import messages from '../messages';
+import messages from '../../../project/general/messages';
 
-const ProjectHeaderImageTooltip = () => {
+const ProjectCardImageTooltip = () => {
   const { formatMessage } = useIntl();
 
   return (
     <IconTooltip
       content={
         <FormattedMessage
-          {...messages.projectHeaderImageTooltip}
+          {...messages.projectCardImageTooltip}
           values={{
             supportPageLink: (
               <a
@@ -30,4 +30,4 @@ const ProjectHeaderImageTooltip = () => {
     />
   );
 };
-export default ProjectHeaderImageTooltip;
+export default ProjectCardImageTooltip;
