+++ front/app/containers/Admin/projects/_shared/components/ProjectSetupForm/ProjectHeaderImageTooltip.tsx
@@ -4,16 +4,16 @@ import { IconTooltip } from '@citizenlab/cl2-component-library';
 
 import { FormattedMessage, useIntl } from 'utils/cl-intl';
 
-import messages from '../messages';
+import messages from '../../../project/general/messages';
 
-const ProjectCardImageTooltip = () => {
+const ProjectHeaderImageTooltip = () => {
   const { formatMessage } = useIntl();
 
   return (
     <IconTooltip
       content={
         <FormattedMessage
-          {...messages.projectCardImageTooltip}
+          {...messages.projectHeaderImageTooltip}
           values={{
             supportPageLink: (
               <a
@@ -30,4 +30,4 @@ const ProjectCardImageTooltip = () => {
     />
   );
 };
-export default ProjectCardImageTooltip;
+export default ProjectHeaderImageTooltip;
