+++ front/app/containers/Admin/projects/_shared/components/ProjectSetupForm/ProjectNameInput.tsx
@@ -10,7 +10,7 @@ import Error from 'components/UI/Error';
 
 import { FormattedMessage } from 'utils/cl-intl';
 
-import messages from '../messages';
+import messages from '../../../project/general/messages';
 
 import { StyledInputMultiloc } from './styling';
