+++ front/app/containers/Admin/projects/_shared/components/ProjectSetupForm/TopicInputsTooltipExtraCopy.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { FormattedMessage } from 'utils/cl-intl';
 import Link from 'utils/cl-router/Link';
 
-import messages from '../messages';
+import messages from '../../../project/general/messages';
 
 const TopicInputsCopyProvider = () => {
   return (
