+++ front/app/containers/Admin/projects/project/proposals/index.tsx
@@ -11,8 +11,8 @@ import ButtonWithLink from 'components/UI/ButtonWithLink';
 
 import { FormattedMessage } from 'utils/cl-intl';
 
-import AnalysisBanner from '../../components/AnalysisBanner';
-import NewIdeaButton from '../../components/NewIdeaButton';
+import AnalysisBanner from '../../_shared/components/AnalysisBanner';
+import NewIdeaButton from '../../_shared/components/NewIdeaButton';
 import messages from '../messages';
 
 type TFilterMenu = 'topics' | 'statuses';
